requirements.txt
# Core ML/RL
torch>=2.0.0
gymnasium>=0.29.0
numpy>=1.24.0

# Data & Logging
pandas>=2.0.0
tensorboard>=2.14.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6

# Testing
pytest>=7.4.0
